serializers split into seperate files, and basis of compliance visit serializer developed
